+++ client/src/components/BulmaEditTable.vue
@@ -1,6 +1,6 @@
 <template>
     <div>
-        <BulmaModal v-if="showEdit" :title="action" action="Save" @click="saveItem()" @close="showEdit=false" @cancel="showEdit=false">
+        <BulmaModal type="large" v-if="showEdit" :title="action" action="Save" @click="saveItem()" @close="showEdit=false" @cancel="showEdit=false">
           <div v-for="field,index in tableFields" :key="field.name" class="field mt-3">
 
               <!-- add field label -->
@@ -19,6 +19,25 @@
                     <option v-for="option in field.values" :key="option" :selected="field.default==option" :value="option">{{ option }}</option>
                   </select>
                 </div>
+                <BulmaAdvancedSelect
+                  v-if="field.type=='query'"
+                  :defaultValue="field.default"
+                  :required="field.required||false"
+                  :multiple="false"
+                  :name="field.name"
+                  :placeholder="field.placeholder||'Select...'"
+                  :values="form[field.from]||[]"
+                  :hasError="$v.editedItem[field.name].$invalid"
+                  :isLoading="!['fixed','variable'].includes(dynamicFieldStatus[field.from])"
+                  v-model="$v.editedItem[field.name].$model"
+                  :icon="field.icon"
+                  :columns="field.columns||[]"
+                  :pctColumns="field.pctColumns||[]"
+                  :previewColumn="field.previewColumn||''"
+                  :valueColumn="field.valueColumn||''"
+                  :sticky="false"
+                  >
+                </BulmaAdvancedSelect>
                 <!-- add left icon, but not for query, because that's a component with icon builtin -->
                 <span v-if="!!field.icon && field.type!='query'" class="icon is-small is-left">
                   <font-awesome-icon :icon="field.icon" />
@@ -92,12 +111,13 @@
     import Vue from 'vue'
     import BulmaModal from './BulmaModal.vue'
     import Vuelidate from 'vuelidate'
+    import BulmaAdvancedSelect from './BulmaAdvancedSelect'
     import { required,minValue,maxValue,minLength,maxLength, helpers,requiredIf } from 'vuelidate/lib/validators'
 
     Vue.use(Vuelidate)
     export default{
         name:'BulmaEditTable',
-        components:{BulmaModal},
+        components:{BulmaModal,BulmaAdvancedSelect},
         props:{
             tableFields: {
                 type: Array,
@@ -112,6 +132,14 @@
                 required: false,
                 type: Boolean,
                 default: false
+            },
+            form:{
+              required: true,
+              type: Object
+            },
+            dynamicFieldStatus:{
+              required: true,
+              type: Object
             }
         },
         data: function(){
@@ -153,7 +181,6 @@
                     {description: description,type:"regex"},
                     (value) => !helpers.req(value) || regexObj.test(value)
                 )
-                console.log(attrs.regex)
               }else{
                 regexObj = new RegExp(ff.regex)
                 description = (ff.regexDescription!==undefined)?ff.regexDescription:"The value must match regular expression : " + ff.regex